+++ crates/holochain/tests/tests/multi_conductor/mod.rs
@@ -1,10 +1,12 @@
 use hdk::prelude::*;
+use std::collections::HashMap;
 //use holochain::conductor::config::{ConductorConfig, DpkiConfig};
 use holochain::sweettest::SweetConductorConfig;
 use holochain::sweettest::*;
 //use holochain_conductor_api::conductor::ConductorTuningParams;
 use holochain_sqlite::db::{DbKindT, DbWrite};
 use holochain_sqlite::prelude::DatabaseResult;
+use holochain_types::network::Kitsune2NetworkMetricsRequest;
 use unwrap_to::unwrap_to;
 
 #[derive(serde::Serialize, serde::Deserialize, Debug, SerializedBytes, derive_more::From)]
@@ -118,7 +120,7 @@ async fn test_publish() -> anyhow::Result<()> {
 async fn multi_conductor() -> anyhow::Result<()> {
     use holochain::test_utils::inline_zomes::simple_create_read_zome;
 
-    holochain_trace::init_fmt(holochain_trace::Output::Log).unwrap();
+    holochain_trace::test_run();
 
     const NUM_CONDUCTORS: usize = 3;
 
@@ -167,21 +169,22 @@ async fn multi_conductor() -> anyhow::Result<()> {
     );
 
     // See if we can fetch metric data from bobbo
-    let metrics = conductors[1].dump_network_metrics(None).await?;
-    tracing::info!(target: "TEST", "@!@! - metrics: {metrics}");
+    let metrics = conductors[1]
+        .dump_network_metrics(Kitsune2NetworkMetricsRequest::default())
+        .await?
+        .into_iter()
+        .map(|(k, v)| (k.to_string(), v))
+        .collect::<HashMap<_, _>>();
+    tracing::info!(target: "TEST", "@!@! - metrics: {}", serde_json::to_string_pretty(&metrics).unwrap());
 
     // See if we can fetch network stats from bobbo
     let stats = conductors[1].dump_network_stats().await?;
-    tracing::info!(target: "TEST", "@!@! - stats: {stats}");
+    tracing::info!(target: "TEST", "@!@! - stats: {}", serde_json::to_string_pretty(&stats).unwrap());
 
-    /*
-    let stats: tx5::stats::Stats = serde_json::from_str(&stats).unwrap();
-
-    // make sure that, by this point, we have upgraded connections to webrtc
-    for con in stats.connection_list {
-        assert!(con.is_webrtc);
-    }
-    */
+    let stats = conductors[1]
+        .dump_network_stats_for_app(&"app".to_string())
+        .await?;
+    tracing::info!(target: "TEST", "@!@! - stats by app: {}", serde_json::to_string_pretty(&stats).unwrap());
 
     Ok(())
 }