+++ crates/holochain_conductor_api/src/admin_interface.rs
@@ -1,4 +1,4 @@
-use std::collections::BTreeSet;
+use std::collections::{BTreeSet, HashMap};
 
 use holo_hash::*;
 use holochain_types::prelude::*;
@@ -275,10 +275,22 @@ pub enum AdminRequest {
     /// [`AdminResponse::NetworkMetricsDumped`]
     DumpNetworkMetrics {
         /// If set, limits the metrics dumped to a single DNA hash space.
+        #[serde(default)]
         dna_hash: Option<DnaHash>,
+
+        /// Whether to include a DHT summary.
+        ///
+        /// You need a dump from multiple nodes in order to make a comparison, so this is not
+        /// requested by default.
+        #[serde(default)]
+        include_dht_summary: bool,
     },
 
-    /// Dump raw json network statistics from the backend networking lib.
+    /// Dump network statistics from the Kitsune2 networking transport module.
+    ///
+    /// # Returns
+    ///
+    /// [`AdminResponse::NetworkStatsDumped`]
     DumpNetworkStats,
 
     /// Add a list of agents to this conductor's peer store.
@@ -535,15 +547,10 @@ pub enum AdminResponse {
     ConductorStateDumped(String),
 
     /// The successful result of a call to [`AdminRequest::DumpNetworkMetrics`].
-    ///
-    /// The string is a JSON blob of the metrics results.
-    NetworkMetricsDumped(String),
+    NetworkMetricsDumped(HashMap<DnaHash, Kitsune2NetworkMetrics>),
 
     /// The successful result of a call to [`AdminRequest::DumpNetworkStats`].
-    ///
-    /// The string is a raw JSON blob returned directly from the backend
-    /// networking library.
-    NetworkStatsDumped(String),
+    NetworkStatsDumped(kitsune2_api::TransportStats),
 
     /// The successful response to an [`AdminRequest::AddAgentInfo`].
     ///