+++ crates/holochain_conductor_api/src/app_interface.rs
@@ -4,6 +4,7 @@ use holochain_keystore::LairResult;
 use holochain_keystore::MetaLairClient;
 use holochain_types::prelude::*;
 use indexmap::IndexMap;
+use std::collections::HashMap;
 
 /// Represents the available conductor functions to call over an app interface
 /// and will result in a corresponding [`AppResponse`] message being sent back over the
@@ -164,12 +165,39 @@ pub enum AppRequest {
     /// [`AppResponse::CloneCellEnabled`]
     EnableCloneCell(Box<EnableCloneCellPayload>),
 
-    /// Info about networking processes
+    /// Retrieve network metrics for the current app.
+    ///
+    /// Identical to what [`AdminRequest::DumpNetworkMetrics`](crate::admin_interface::AdminRequest::DumpNetworkMetrics)
+    /// does, but scoped to the current app.
+    ///
+    /// If `dna_hash` is not set, metrics for all DNAs in the current app are returned.
     ///
     /// # Returns
     ///
-    /// [`AppResponse::NetworkInfo`]
-    NetworkInfo(Box<NetworkInfoRequestPayload>),
+    /// [`AppResponse::NetworkMetricsDumped`]
+    DumpNetworkMetrics {
+        /// If set, limits the metrics dumped to a single DNA hash space.
+        #[serde(default)]
+        dna_hash: Option<DnaHash>,
+
+        /// Whether to include a DHT summary.
+        ///
+        /// You need a dump from multiple nodes in order to make a comparison, so this is not
+        /// requested by default.
+        #[serde(default)]
+        include_dht_summary: bool,
+    },
+
+    /// Dump network statistics from the Kitsune2 networking transport module.
+    ///
+    /// Identical to what [`AdminRequest::DumpNetworkStats`](crate::admin_interface::AdminRequest::DumpNetworkStats)
+    /// does, but scoped to the current app. Connections that are not relevant to a DNA in the
+    /// current app are filtered out.
+    ///
+    /// # Returns
+    ///
+    /// [`AppResponse::NetworkStatsDumped`]
+    DumpNetworkStats,
 
     /// List all host functions available to wasm on this conductor.
     ///
@@ -261,8 +289,11 @@ pub enum AppResponse {
     /// is returned.
     CloneCellEnabled(ClonedCell),
 
-    /// NetworkInfo is returned
-    NetworkInfo(Vec<NetworkInfo>),
+    /// The successful result of a call to [`AppRequest::DumpNetworkMetrics`].
+    NetworkMetricsDumped(HashMap<DnaHash, Kitsune2NetworkMetrics>),
+
+    /// The successful result of a call to [`AppRequest::DumpNetworkStats`].
+    NetworkStatsDumped(kitsune2_api::TransportStats),
 
     /// All the wasm host functions supported by this conductor.
     ListWasmHostFunctions(Vec<String>),
@@ -530,27 +561,6 @@ impl From<AppInfoStatus> for AppStatus {
             AppInfoStatus::AwaitingMemproofs => AppStatus::AwaitingMemproofs,
         }
     }
-}
-
-/// Temporarily copying this here until we actually implement something
-/// equivalent in kitsune2.
-#[derive(Clone, Debug, Default, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub struct FetchPoolInfo {
-    /// Total number of bytes expected to be received through fetches
-    pub op_bytes_to_fetch: usize,
-
-    /// Total number of ops expected to be received through fetches
-    pub num_ops_to_fetch: usize,
-}
-
-#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize, SerializedBytes)]
-pub struct NetworkInfo {
-    pub fetch_pool_info: FetchPoolInfo,
-    pub current_number_of_peers: u32,
-    pub arc_size: f64,
-    pub total_network_peers: u32,
-    pub bytes_since_last_time_queried: u64,
-    pub completed_rounds_since_last_time_queried: u32,
 }
 
 /// The request payload sent on a Holochain app websocket to authenticate the connection.