+++ crates/holochain_sqlite/src/sql/cell/sum_of_received_bytes_since_timestamp.sql
@@ -1,8 +1,0 @@
-SELECT
-  SUM(LENGTH(Action.blob))
-FROM
-  Action,
-  DhtOp
-WHERE
-  DhtOp.authored_timestamp > ?1
-  AND DhtOp.action_hash = Action.hash;