+++ crates/holochain_terminal/src/components/network_metrics.rs
@@ -3,18 +3,20 @@ use crate::client::AppClient;
 use crate::components::common::show_message;
 use crate::event::ScreenEvent;
 use anyhow::anyhow;
+use chrono::{DateTime, Utc};
 use holo_hash::{AgentPubKey, DnaHash};
-use holochain_conductor_api::NetworkInfo;
+use holochain_types::network::Kitsune2NetworkMetrics;
 use holochain_util::tokio_helper::block_on;
 use once_cell::sync::Lazy;
 use ratatui::{prelude::*, widgets::*};
+use std::collections::HashMap;
 use std::sync::{Arc, OnceLock, RwLock};
 use std::time::{Duration, Instant};
 use tokio::sync::Mutex;
 
 type NetworkInfoParams = (AgentPubKey, Vec<(String, DnaHash)>);
 
-static NETWORK_INFO_PARAMS: Lazy<Arc<RwLock<Option<NetworkInfoParams>>>> =
+static NETWORK_METRICS_PARAMS: Lazy<Arc<RwLock<Option<NetworkInfoParams>>>> =
     Lazy::new(|| Arc::new(RwLock::new(None)));
 
 fn get_selected() -> &'static RwLock<usize> {
@@ -29,19 +31,20 @@ fn get_last_refresh_at() -> &'static RwLock<Option<Instant>> {
     LAST_REFRESH_AT.get_or_init(|| RwLock::new(None))
 }
 
-fn get_network_info() -> &'static RwLock<Vec<NetworkInfo>> {
-    static NETWORK_INFO: OnceLock<RwLock<Vec<NetworkInfo>>> = OnceLock::new();
+fn get_network_metrics() -> &'static RwLock<HashMap<DnaHash, Kitsune2NetworkMetrics>> {
+    static NETWORK_INFO: OnceLock<RwLock<HashMap<DnaHash, Kitsune2NetworkMetrics>>> =
+        OnceLock::new();
 
-    NETWORK_INFO.get_or_init(|| RwLock::new(vec![]))
+    NETWORK_INFO.get_or_init(|| RwLock::new(HashMap::with_capacity(0)))
 }
 
-pub struct NetworkInfoWidget {
+pub struct NetworkMetricsWidget {
     args: Arc<Args>,
     app_client: Option<Arc<Mutex<AppClient>>>,
     events: Vec<ScreenEvent>,
 }
 
-impl NetworkInfoWidget {
+impl NetworkMetricsWidget {
     pub fn new(
         args: Arc<Args>,
         app_client: Option<Arc<Mutex<AppClient>>>,
@@ -55,7 +58,7 @@ impl NetworkInfoWidget {
     }
 }
 
-impl Widget for NetworkInfoWidget {
+impl Widget for NetworkMetricsWidget {
     fn render(self, area: Rect, buf: &mut Buffer) {
         let app_id = match &self.args.app_id {
             Some(b) => b.clone(),
@@ -73,9 +76,9 @@ impl Widget for NetworkInfoWidget {
             }
         };
 
-        if NETWORK_INFO_PARAMS.read().unwrap().is_none() {
-            *NETWORK_INFO_PARAMS.write().unwrap() =
-                match get_network_info_params(app_client.clone(), app_id.clone()) {
+        if NETWORK_METRICS_PARAMS.read().unwrap().is_none() {
+            *NETWORK_METRICS_PARAMS.write().unwrap() =
+                match get_network_metrics_params(app_client.clone(), app_id.clone()) {
                     Ok(p) => Some(p),
                     Err(e) => {
                         show_message(format!("{:?}", e).as_str(), area, buf);
@@ -98,9 +101,9 @@ impl Widget for NetworkInfoWidget {
             }
 
             if do_refresh {
-                match fetch_network_info(app_client) {
-                    Ok(network_infos) => {
-                        *get_network_info().write().unwrap() = network_infos;
+                match fetch_network_metrics(app_client) {
+                    Ok(metrics) => {
+                        *get_network_metrics().write().unwrap() = metrics;
                     }
                     Err(e) => {
                         show_message(format!("{:?}", e).as_str(), area, buf);
@@ -110,12 +113,16 @@ impl Widget for NetworkInfoWidget {
             };
         }
 
-        let network_infos = get_network_info().read().unwrap();
+        let network_metrics_params = NETWORK_METRICS_PARAMS.read().unwrap();
+        let network_metrics_params_value = network_metrics_params
+            .as_ref()
+            .expect("Should have network metrics params");
+
         for event in self.events {
             match event {
                 ScreenEvent::NavDown => {
                     let mut selected = get_selected().write().unwrap();
-                    if *selected < network_infos.len() - 1 {
+                    if *selected < network_metrics_params_value.1.len() - 1 {
                         *selected += 1;
                     }
                 }
@@ -137,7 +144,7 @@ impl Widget for NetworkInfoWidget {
             .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
             .split(area);
 
-        let list_items: Vec<ListItem> = NETWORK_INFO_PARAMS
+        let list_items: Vec<ListItem> = NETWORK_METRICS_PARAMS
             .read()
             .unwrap()
             .clone()
@@ -166,27 +173,86 @@ impl Widget for NetworkInfoWidget {
             .highlight_symbol(">> ");
 
         let selected = *get_selected().read().unwrap();
-        let selected = if !network_infos.is_empty() && selected < network_infos.len() {
+
+        let selected = if !network_metrics_params_value.1.is_empty()
+            && selected < network_metrics_params_value.1.len()
+        {
+            let dna_hash = &network_metrics_params_value.1[selected].1;
+
+            let network_metrics = get_network_metrics().read().unwrap();
+            let Some(metrics) = network_metrics.get(dna_hash) else {
+                return;
+            };
+
             let detail_line = List::new(vec![
+                // Fetch
                 ListItem::new(format!(
-                    "peers             : {:?}",
-                    network_infos[selected].current_number_of_peers,
+                    "Pending fetch requests : {:?}",
+                    metrics.fetch_state_summary.pending_requests.len(),
                 )),
                 ListItem::new(format!(
-                    "total peers       : {:?}",
-                    network_infos[selected].total_network_peers
+                    "Fetch peers on backoff : {:?}",
+                    metrics.fetch_state_summary.peers_on_backoff.len(),
+                )),
+                // Gossip
+                ListItem::new(format!(
+                    "Gossip peers           : {:?}",
+                    metrics.gossip_state_summary.peer_meta.len(),
                 )),
                 ListItem::new(format!(
-                    "total bytes       : {:?}",
-                    network_infos[selected].bytes_since_last_time_queried
+                    "Last gossip round      : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.last_gossip_timestamp)
+                        .max()
+                        .map(|ts| { DateTime::<Utc>::from_timestamp(ts.as_micros() / 1000, 0) }),
                 )),
                 ListItem::new(format!(
-                    "num ops to fetch  : {:?}",
-                    network_infos[selected].fetch_pool_info.num_ops_to_fetch
+                    "Peer behaviour errors  : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.peer_behavior_errors)
+                        .sum::<u32>(),
                 )),
                 ListItem::new(format!(
-                    "op bytes to fetch : {:?}",
-                    network_infos[selected].fetch_pool_info.op_bytes_to_fetch
+                    "Local errors           : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.local_errors)
+                        .sum::<u32>(),
+                )),
+                ListItem::new(format!(
+                    "Peer busy              : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.peer_busy)
+                        .sum::<u32>(),
+                )),
+                ListItem::new(format!(
+                    "Completed rounds       : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.completed_rounds)
+                        .sum::<u32>(),
+                )),
+                ListItem::new(format!(
+                    "Peer timeouts          : {:?}",
+                    metrics
+                        .gossip_state_summary
+                        .peer_meta
+                        .iter()
+                        .filter_map(|meta| meta.1.peer_timeouts)
+                        .sum::<u32>(),
                 )),
             ])
             .block(Block::default().title(" Info ").borders(Borders::ALL))
@@ -208,7 +274,7 @@ impl Widget for NetworkInfoWidget {
     }
 }
 
-fn get_network_info_params(
+fn get_network_metrics_params(
     app_client: Arc<Mutex<AppClient>>,
     app_id: String,
 ) -> anyhow::Result<(AgentPubKey, Vec<(String, DnaHash)>)> {
@@ -217,34 +283,26 @@ fn get_network_info_params(
             app_client
                 .lock()
                 .await
-                .discover_network_info_params(app_id)
+                .discover_network_metrics_params(app_id)
                 .await
         },
         Duration::from_secs(10),
     ) {
         Ok(Ok(p)) => Ok(p),
-        Ok(Err(e)) => Err(anyhow!("Error fetching network info params - {:?}", e)),
-        Err(_) => Err(anyhow!("Timeout while fetching network info params")),
+        Ok(Err(e)) => Err(anyhow!("Error fetching network metrics params - {:?}", e)),
+        Err(_) => Err(anyhow!("Timeout while fetching network metrics params")),
     }
 }
 
-fn fetch_network_info(app_client: Arc<Mutex<AppClient>>) -> anyhow::Result<Vec<NetworkInfo>> {
-    let (agent, named_dna_hashes) = NETWORK_INFO_PARAMS.read().unwrap().clone().unwrap();
+fn fetch_network_metrics(
+    app_client: Arc<Mutex<AppClient>>,
+) -> anyhow::Result<HashMap<DnaHash, Kitsune2NetworkMetrics>> {
     match block_on(
-        async {
-            app_client
-                .lock()
-                .await
-                .network_info(
-                    agent,
-                    named_dna_hashes.into_iter().map(|(_, h)| h).collect(),
-                )
-                .await
-        },
+        async { app_client.lock().await.network_metrics().await },
         Duration::from_secs(10),
     ) {
-        Ok(Ok(network_infos)) => Ok(network_infos),
-        Ok(Err(e)) => Err(anyhow!("Failed to fetch network infos - {:?}", e)),
-        Err(_) => Err(anyhow!("Timeout while fetching network infos")),
+        Ok(Ok(metrics)) => Ok(metrics),
+        Ok(Err(e)) => Err(anyhow!("Failed to fetch network metrics - {:?}", e)),
+        Err(_) => Err(anyhow!("Timeout while fetching network metrics")),
     }
 }